// server/src/types.ts
export interface User {
  id: number;
  avatar: string;
  first_name: string;
  last_name: string;
  age: number;
  nationality: string;
  hobbies: string[];
}

export interface FilterCount {
  value: string;
  count: number;
}

export interface UsersResponse {
  users: User[];
  nextCursor: string | null;
  hasMore: boolean;
  topHobbies: FilterCount[];
  topNationalities: FilterCount[];
}